Fix duplicated module header in app factory
