+++ src/app/main.py
@@ -3,8 +3,14 @@ from __future__ import annotations
 
 from fastapi import FastAPI, File, HTTPException, UploadFile, status
 
-from app.application.process_document import ProcessDocumentUseCase, RetrieveDocumentUseCase
+from app.application.process_document import (
+    DocumentParser,
+    ProcessDocumentUseCase,
+    RetrieveDocumentUseCase,
+)
 from app.config.settings import get_settings
+from app.domain.repositories.classification_repository import ClassificationRepository
+from app.domain.repositories.document_repository import DocumentRepository
 from app.domain.services.classification_extractor import extract_classification
 from app.infrastructure.parsers.pdf_document_parser import PdfDocumentParser
 from app.infrastructure.repositories.json_classification_repository import (
@@ -12,15 +18,78 @@ from app.infrastructure.repositories.json_classification_repository import (
 )
 from app.infrastructure.repositories.json_file_repository import JsonFileRepository
 
-app = FastAPI(title="Document Processor API", version="0.1.0")
 
-settings = get_settings()
-parser = PdfDocumentParser()
-classification_repository = JsonClassificationRepository(settings.classification_path)
+def create_app(
+    document_parser: DocumentParser | None = None,
+    classification_repo: ClassificationRepository | None = None,
+    schedule_repo: DocumentRepository | None = None,
+) -> FastAPI:
+    """Create and configure the FastAPI application instance."""
 
-schedule_repository = JsonFileRepository(settings.schedule_path)
-schedule_processor = ProcessDocumentUseCase(parser, schedule_repository)
-schedule_retriever = RetrieveDocumentUseCase(schedule_repository)
+    settings = get_settings()
+    pdf_parser = document_parser or PdfDocumentParser()
+    classification_repository = (
+        classification_repo
+        or JsonClassificationRepository(settings.classification_path)
+    )
+    schedule_repository = schedule_repo or JsonFileRepository(settings.schedule_path)
+    schedule_processor = ProcessDocumentUseCase(pdf_parser, schedule_repository)
+    schedule_retriever = RetrieveDocumentUseCase(schedule_repository)
+
+    app = FastAPI(title="Document Processor API", version="0.1.0")
+
+    @app.post("/classification/pdf", status_code=status.HTTP_201_CREATED)
+    async def upload_classification(file: UploadFile = File(...)) -> dict:
+        """Parse and persist the uploaded classification PDF, returning its JSON form."""
+
+        pdf_bytes = await _read_pdf_bytes(file)
+        parsed_document = pdf_parser.parse(pdf_bytes)
+        try:
+            classification_table = extract_classification(parsed_document)
+        except ValueError as extraction_error:
+            raise HTTPException(
+                status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
+                detail=str(extraction_error),
+            ) from extraction_error
+        classification_repository.save(classification_table)
+        return classification_table.to_dict()
+
+    @app.get("/classification", status_code=status.HTTP_200_OK)
+    async def get_classification() -> dict:
+        """Retrieve the stored classification document as JSON."""
+
+        classification_table = classification_repository.load()
+        if classification_table is None:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail="No processed classification document available.",
+            )
+        return classification_table.to_dict()
+
+    @app.post("/schedule/pdf", status_code=status.HTTP_201_CREATED)
+    async def upload_schedule(file: UploadFile = File(...)) -> dict:
+        """Parse and persist the uploaded schedule PDF, returning its JSON form."""
+
+        pdf_bytes = await _read_pdf_bytes(file)
+        parsed_document = schedule_processor.execute(pdf_bytes)
+        return parsed_document.to_dict()
+
+    @app.get("/schedule", status_code=status.HTTP_200_OK)
+    async def get_schedule() -> dict:
+        """Retrieve the stored schedule document as JSON."""
+
+        parsed_document = schedule_retriever.execute()
+        if parsed_document is None:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail="No processed schedule document available.",
+            )
+        return parsed_document.to_dict()
+
+    return app
+
+
+app = create_app()
 
 
 async def _read_pdf_bytes(uploaded_file: UploadFile) -> bytes:
@@ -37,54 +106,6 @@ async def _read_pdf_bytes(uploaded_file: UploadFile) -> bytes:
             detail="The PDF file is empty.",
         )
     return file_bytes
-
-
-@app.post("/classification/pdf", status_code=status.HTTP_201_CREATED)
-async def upload_classification(file: UploadFile = File(...)) -> dict:
-    """Parse and persist the uploaded classification PDF, returning its JSON form."""
-    pdf_bytes = await _read_pdf_bytes(file)
-    parsed_document = parser.parse(pdf_bytes)
-    try:
-        classification_table = extract_classification(parsed_document)
-    except ValueError as extraction_error:
-        raise HTTPException(
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            detail=str(extraction_error),
-        ) from extraction_error
-    classification_repository.save(classification_table)
-    return classification_table.to_dict()
-
-
-@app.get("/classification", status_code=status.HTTP_200_OK)
-async def get_classification() -> dict:
-    """Retrieve the stored classification document as JSON."""
-    classification_table = classification_repository.load()
-    if classification_table is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="No processed classification document available.",
-        )
-    return classification_table.to_dict()
-
-
-@app.post("/schedule/pdf", status_code=status.HTTP_201_CREATED)
-async def upload_schedule(file: UploadFile = File(...)) -> dict:
-    """Parse and persist the uploaded schedule PDF, returning its JSON form."""
-    pdf_bytes = await _read_pdf_bytes(file)
-    parsed_document = schedule_processor.execute(pdf_bytes)
-    return parsed_document.to_dict()
-
-
-@app.get("/schedule", status_code=status.HTTP_200_OK)
-async def get_schedule() -> dict:
-    """Retrieve the stored schedule document as JSON."""
-    parsed_document = schedule_retriever.execute()
-    if parsed_document is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="No processed schedule document available.",
-        )
-    return parsed_document.to_dict()
 
 
 if __name__ == "__main__":
